pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "adminos-lab"
version = "0.1.0"
description = "AdminOS Lab Django project"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "Django==5.2",
    "django-environ>=0.11",
    "psycopg[binary]>=3.1",
    "weasyprint>=61.0",
    "reportlab>=4.0",
]

[tool.django]
settings-module = "adminos_lab.settings"

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "adminos_lab.settings"
python_files = ["tests.py", "test_*.py", "*_tests.py"]
